renames db utils test suite and data for carity and puts tests in classes with custom markers
